hand_recognition/requirements.txt
addict          == 2.2.1      
aniso8601       == 9.0.1      
certifi         == 2020.6.20  
chardet         == 3.0.4      
click           == 7.1.2      
cycler          == 0.10.0     
Cython          == 0.29.21         
idna            == 2.6        
itsdangerous    == 1.1.0      
Jinja2          == 2.11.2     
kazoo           == 2.5.0      
kiwisolver      == 1.2.0      
MarkupSafe      == 1.1.1      
matplotlib      == 3.3.0      
mmcv            == 0.6.2      
numpy           == 1.19.1     
opencv-python   == 4.2.0.34   
Pillow          == 6.2.2      
pip             == 10.0.1      
pycocotools     == 2.0.1      
pykafka         == 2.8.0      
pyparsing       == 2.4.7      
python-dateutil == 2.8.1      
pytz            == 2022.7.1   
PyYAML          == 5.3.1      
redis           == 3.4.1      
requests        == 2.18.4     
setuptools      == 39.0.1     
Shapely         == 1.7.0      
six             == 1.15.0     
tabulate        == 0.9.0      
terminaltables  == 3.1.0      
torch           == 1.4.0
torchvision     == 0.5.0
urllib3         == 1.22       
Werkzeug        == 1.0.1      
yapf            == 0.30.0 